【AC】[CODE FESTIVAL 2017 qual B]C - 3 Steps

+++ code-festival-2017-qualb/code_festival_2017_qualb_c.cpp
@@ -106,7 +106,7 @@ int main() {
         ret();
     }
 
-    ll all = n * (n - 1) / 2;
+    ll all = ll(n) * (n - 1) / 2;
     ll ans = all - m;
     cout << ans << endl;
 }
